feat: render operable slave email provider
was able to render an operable send grid service that sends emails
to a users listed contacts. Unfortunately, I haven't been able to send
emails via sparkpost without having to pay for a DNS, which I'm not
ready to do at this point.  Perhaps in the future.  As of now, will
focus now on writing documentation, and a simple UI, that can issue
commands via the api that has been built.